+++ src/components/GifGrid.jsx
@@ -1,28 +1,22 @@
-import { useEffect, useState } from 'react'
-import { getGif } from '../helpers/getGifs'
 import { GifItem } from './GifItem'
+import { useFetchGifs } from '../hook/useFetchGifs'
 
 // eslint-disable-next-line react/prop-types
 export function GifGrid ({ category }) {
-  const [image, setImage] = useState([])
-
-  useEffect(() => {
-    const newImage = async () => {
-      const newImg = await getGif(category)
-      setImage(newImg)
-    }
-
-    newImage()
-  }, [category])
+  const { image, isLoading } = useFetchGifs({ category })
 
   return (
     <>
       <h4>{category}</h4>
       <div className='card-grid'>
         {
-          image?.map((img) => (
-            <GifItem key={img.id} {...img} />
-          ))
+          isLoading
+            ? <div>Cargando...</div>
+            : (
+                image?.map((img) => (
+                  <GifItem key={img.id} {...img} />
+                ))
+              )
         }
       </div>
     </>
